small fixes

+++ textdistance/algorithms/base.py
@@ -155,7 +155,7 @@ class Base:
             result += s
         return result
 
-    def _count_counters(self, counter: Counter) -> float:
+    def _count_counters(self, counter: Counter) -> int:
         """Return all elements count from Counter
         """
         if getattr(self, 'as_set', False):
